feat(storybook): add AppDecorator to wrap stories with AppProvider for consistent context fix(storybook): update stories path in main.ts to include all .stories.tsx files refactor(page): reorganize component imports to reflect new directory structure feat(stories): create page.stories.tsx for Home page to facilitate Storybook integration fix(tests): update import path for mantineTheme in index.spec.tsx to match new structure
feat(AppProvider): update import path for mantineTheme to improve structure
feat(ColorSchemeToggle): add ColorSchemeToggle component for theme switching
feat(Faq): create Faq component with styles and stories for better documentation and testing

feat(Faq): add FAQ component to provide users with answers to common questions
feat(Features.module.css): create styles for FAQ component and other UI elements to enhance user experience

feat(Features): add Features component and its Storybook stories to showcase key functionalities and improve documentation

feat(footer): add Footer component with social media links and styles for better user engagement
feat(footer): create Footer stories for Storybook to facilitate component testing and documentation

feat(Header): add Header component with styles and Storybook integration for better UI representation and testing

feat(Hero): add Hero component styles and Storybook configuration to enhance UI and documentation

feat(client): add Hero component for showcasing AIGraphr features and create mantine theme configuration
chore(client): remove empty .gitkeep file from components stories directory
