+++ apps/client/.storybook/preview.ts
@@ -1,6 +1,6 @@
 import {ReactRenderer} from '@storybook/react';
 import {DecoratorFunction} from '@storybook/types';
-import {AppDecorator} from '../stories/decorators/AppDecorator';
+import {AppDecorator} from './decorators/AppDecorator';
 import './../app/global.css';
 
 export const decorators: DecoratorFunction<ReactRenderer>[] = [AppDecorator()];
