+++ apps/client/src/components/home/Faq.stories.tsx
@@ -1,0 +1,14 @@
+import {Meta, StoryObj} from '@storybook/react';
+import {Faq as Component} from './Faq';
+
+const meta = {
+    title: 'Components/Home/Faq',
+    component: Component,
+    args: {}
+} satisfies Meta<typeof Component>;
+
+export default meta;
+
+export const Faq = {
+    render: (props) => <Component {...props} />
+} satisfies StoryObj<typeof meta>;
