+++ apps/client/src/components/home/Faq.tsx
