+++ apps/client/src/components/home/Footer.tsx
