+++ apps/client/src/components/home/Hero.stories.tsx
@@ -1,0 +1,14 @@
+import {Meta, StoryObj} from '@storybook/react';
+import {Hero as Component} from './Hero';
+
+const meta = {
+    title: 'Components/Home/Hero',
+    component: Component,
+    args: {}
+} satisfies Meta<typeof Component>;
+
+export default meta;
+
+export const Hero = {
+    render: (props) => <Component {...props} />
+} satisfies StoryObj<typeof meta>;
