+++ apps/client/src/components/home/Hero.tsx
